initial commit-- swapping stations for db secrets
- nav
- about
- contact
- GET count-emails by rep
- cleanup of root route `/`
