+++ app/about/page.tsx
@@ -1,0 +1,7 @@
+export default function AboutPage() {
+    return (
+        <main>
+            about page stub
+        </main>
+    );
+};
