+++ app/api/submit-form/route.ts
@@ -1,6 +1,6 @@
 import { NextResponse } from 'next/server';
 import { contactSchiff } from '@/utils/fillOutForms';
-import postMango from '@/utils/mango';
+import { connectAndPost as postMango } from '@/utils/mango';
 import { normalizePhoneNumber } from '@/utils/beCommonUtils';
 
 export async function POST(request: Request) {
