+++ app/contact/page.tsx
@@ -1,0 +1,10 @@
+export default function ContactPage() {
+    return (
+        <main>
+            <h1>Contact Page</h1>
+
+            <p>Hi, welcome to the Contact Page. If you need to contact the owner of this page please reach out to <strong>thomkaar@gmail.com</strong></p>
+            <p>If you would like to help contribute feel free open a branch and a PR at: https://github.com/ThomKaar/ca-for-palestine </p>
+        </main>
+    );
+};
