+++ app/utils/beCommonUtils.ts
@@ -1,4 +1,5 @@
 const DEFAULT_NUMBER = '18054416665';
+export const REPS = ['Schiff', 'Padilla'];
 
 export const normalizePhoneNumber = (s: string) => {
     const onlyNums = s.split('').filter((e) => !isNaN(parseInt(e)));
